+++ packages/utils/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [4.5.0](https://github.com/newsuk/times-components/compare/@times-components/utils@4.4.2...@times-components/utils@4.5.0) (2019-02-18)
+
+
+### Features
+
+* Article Magazine Standard tablet byline ([#1709](https://github.com/newsuk/times-components/issues/1709)) ([904cd23](https://github.com/newsuk/times-components/commit/904cd23))
+
+
+
+
+
 ## [4.4.2](https://github.com/newsuk/times-components/compare/@times-components/utils@4.4.1...@times-components/utils@4.4.2) (2019-02-18)
 
 **Note:** Version bump only for package @times-components/utils